Fix check old streams in metadata, download manager worker thread path

+++ server/StreamManager.js
@@ -65,12 +65,13 @@ class StreamManager {
       if (!dirs || !dirs.length) return true
 
       await Promise.all(dirs.map(async (dirname) => {
-        if (dirname !== 'streams' && dirname !== 'books') {
+        if (dirname !== 'streams' && dirname !== 'books' && dirname !== 'downloads') {
           var fullPath = Path.join(this.MetadataPath, dirname)
           Logger.warn(`Removing OLD Orphan Stream ${dirname}`)
           return fs.remove(fullPath)
         }
       }))
+
       return true
     } catch (error) {
       Logger.debug('No old orphan streams', error)
